+++ src/search/RandomSearchAlgorithm.cpp
@@ -65,7 +65,7 @@ PartitioningScheme * RandomSearchAlgorithm::start(
 			/* building first scheme */
 			t_partitioningScheme * firstSchemeId = new t_partitioningScheme(
 					number_of_genes);
-			for (unsigned int gene = 0; gene < number_of_genes; gene++) {
+			for (size_t gene = 0; gene < number_of_genes; gene++) {
 				t_partitionElementId geneId(1);
 				geneId.at(0) = gene;
 				firstSchemeId->at(gene) = geneId;
@@ -99,7 +99,7 @@ PartitioningScheme * RandomSearchAlgorithm::start(
 				delete bestScheme;
 				bestScheme = localBestScheme;
 				PartitionMap::getInstance()->keep(bestScheme->getId());
-				for (unsigned int i = 0;
+				for (size_t i = 0;
 						i < localBestScheme->getNumberOfElements(); i++) {
 					PartitionMap::getInstance()->purgePartitionMap(
 							localBestScheme->getElement(i)->getId());
